+++ samples/amqsconntls.go
@@ -1,25 +1,22 @@
 /*
 This is a short sample to show how to connect to a remote
-queue manager in a Go program without requiring external
-client configuration such as a CCDT. Only the basic
-parameters are needed here - channel name and connection information -
+queue manager in a Go program with a TLS connection without requiring external
+client configuration such as a CCDT. The basic
+parameters are still needed here - channel name and connection information -
 along with the queue manager name.
 
 For example, run as
-   amqsconn QMGR1 "SYSTEM.DEF.SVRCONN" "myhost.example.com(1414)"
+   amqsconn QMGR1 "SYSTEM.SSL.SVRCONN" "myhost.example.com(1414)"
 
 If the MQSAMP_USER_ID environment variable is set, then a userid/password
 flow is also made to authenticate to the queue manager.
-
-There is no attempt in this sample to configure advanced security features
-such TLS.
 
 If an error occurs, the error is reported.
 */
 package main
 
 /*
-  Copyright (c) IBM Corporation 2017, 2018
+  Copyright (c) IBM Corporation 2017, 2019
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -53,10 +50,10 @@ func main() {
 	var rc int
 
 	if len(os.Args) != 4 {
-		fmt.Println("amqsconn <qmgrname> <channelname> <conname>")
+		fmt.Println("amqsconntls <qmgrname> <channelname> <conname>")
 		fmt.Println("")
 		fmt.Println("For example")
-		fmt.Println("  amqsconn QMGR1 \"SYSTEM.DEF.SVRCONN\" \"myhost.example.com(1414)\"")
+		fmt.Println("  amqsconntls QMGR1 \"SYSTEM.TLS.SVRCONN\" \"myhost.example.com(1414)\"")
 		fmt.Println("All parameters are required.")
 		os.Exit(1)
 	}
@@ -66,20 +63,31 @@ func main() {
 
 	// Allocate the MQCNO and MQCD structures needed for the CONNX call.
 	cno := ibmmq.NewMQCNO()
+	sco := ibmmq.NewMQSCO()
 	cd := ibmmq.NewMQCD()
 
 	// Fill in required fields in the MQCD channel definition structure
 	cd.ChannelName = os.Args[2]
 	cd.ConnectionName = os.Args[3]
 
+	// The CipherSpec must match what is configured on the corresponding SVRCONN
+	cd.SSLCipherSpec = "TLS_RSA_WITH_AES_128_CBC_SHA256"
+
+	// The ClientAuth field says whether or not the client needs to present its own certificate
+	// This too must match the SVRCONN definition.
+	cd.SSLClientAuth = ibmmq.MQSCA_OPTIONAL
+
+	// The keystore contains at least the certificate to verify the qmgr's cert (usually from
+	// a Certificate Authority) and optionally the client's own certificate.
+	// We could also optionally specify which certificate represents the client, based on its label
+	// but don't need to do this when using the MQSCA_OPTIONAL flag.
+	sco.KeyRepository = "./mykey"
+
 	// Reference the CD structure from the CNO and indicate that we definitely want to
 	// use the client connection method.
 	cno.ClientConn = cd
 	cno.Options = ibmmq.MQCNO_CLIENT_BINDING
-
-	// MQ V9.1.2 allows applications to specify their own names. This is ignored
-	// by older levels of the MQ libraries.
-	cno.ApplName = "Golang 9.1.2 ApplName"
+	cno.SSLConfig = sco
 
 	// Also fill in the userid and password if the MQSAMP_USER_ID
 	// environment variable is set. This is the same variable used by the C